# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Learning Management Service",
    "description": "A learning management service API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/interactions/": {
      "get": {
        "tags": [
          "interactions"
        ],
        "summary": "Get Interactions",
        "description": "Get all interactions, optionally filtered by item.",
        "operationId": "get_interactions_interactions__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InteractionModel"
                  },
                  "title": "Response Get Interactions Interactions  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "interactions"
        ],
        "summary": "Post Interaction",
        "description": "Create a new interaction log.",
        "operationId": "post_interaction_interactions__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionLogCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionLog"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/analytics/scores": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Scores",
        "description": "Score distribution histogram for a given lab.",
        "operationId": "get_scores_analytics_scores_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lab",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Lab identifier, e.g. 'lab-01'",
              "title": "Lab"
            },
            "description": "Lab identifier, e.g. 'lab-01'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/pass-rates": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Pass Rates",
        "description": "Per-task pass rates for a given lab.",
        "operationId": "get_pass_rates_analytics_pass_rates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lab",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Lab identifier, e.g. 'lab-01'",
              "title": "Lab"
            },
            "description": "Lab identifier, e.g. 'lab-01'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "integer"
                        }
                      ]
                    }
                  },
                  "title": "Response Get Pass Rates Analytics Pass Rates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/timeline": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Timeline",
        "description": "Submissions per day for a given lab.",
        "operationId": "get_timeline_analytics_timeline_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lab",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Lab identifier, e.g. 'lab-01'",
              "title": "Lab"
            },
            "description": "Lab identifier, e.g. 'lab-01'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "integer"
                        }
                      ]
                    }
                  },
                  "title": "Response Get Timeline Analytics Timeline Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/groups": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Groups",
        "description": "Per-group performance for a given lab.",
        "operationId": "get_groups_analytics_groups_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lab",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Lab identifier, e.g. 'lab-01'",
              "title": "Lab"
            },
            "description": "Lab identifier, e.g. 'lab-01'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "integer"
                        }
                      ]
                    }
                  },
                  "title": "Response Get Groups Analytics Groups Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/completion-rate": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Completion Rate",
        "description": "Completion rate for a given lab \u2014 percentage of learners who scored >= 60.",
        "operationId": "get_completion_rate_analytics_completion_rate_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lab",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Lab identifier, e.g. 'lab-01'",
              "title": "Lab"
            },
            "description": "Lab identifier, e.g. 'lab-01'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/top-learners": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Top Learners",
        "description": "Top learners by average score for a given lab.",
        "operationId": "get_top_learners_analytics_top_learners_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lab",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Lab identifier, e.g. 'lab-01'",
              "title": "Lab"
            },
            "description": "Lab identifier, e.g. 'lab-01'"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of top learners to return",
              "default": 10,
              "title": "Limit"
            },
            "description": "Number of top learners to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "number"
                        }
                      ]
                    }
                  },
                  "title": "Response Get Top Learners Analytics Top Learners Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InteractionLog": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "learner_id": {
            "type": "integer",
            "title": "Learner Id"
          },
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "checks_passed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checks Passed"
          },
          "checks_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checks Total"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "learner_id",
          "item_id",
          "kind"
        ],
        "title": "InteractionLog",
        "description": "An interaction log entry in the database."
      },
      "InteractionLogCreate": {
        "properties": {
          "learner_id": {
            "type": "integer",
            "title": "Learner Id"
          },
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "learner_id",
          "item_id",
          "kind"
        ],
        "title": "InteractionLogCreate",
        "description": "Request schema for creating an interaction log.",
        "example": "{\n                \"learner_id\": 1,\n                \"item_id\": 1,\n                \"kind\": \"attempt\"\n            }"
      },
      "InteractionModel": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "learner_id": {
            "type": "integer",
            "title": "Learner Id"
          },
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "learner_id",
          "item_id",
          "kind",
          "created_at"
        ],
        "title": "InteractionModel",
        "description": "Response schema for an interaction."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}